Refactor into lib
