+++ src/bitslideslib/syncjob.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, PartialEq)]
+pub struct SyncJob {
+    pub src: String,
+    pub dst: String,
+    pub issue: String,
+}
