+++ src/lib.rs
@@ -1,0 +1,5 @@
+mod bitslideslib;
+
+pub use bitslideslib::slide;
+pub use bitslideslib::config::GlobalConfig;
+pub use bitslideslib::config::RootsetConfig;
